DF-10193 fix after merge development
Former-commit-id: b8f6091b8928b4440a5b3630a5847aa76757442d

+++ Sources/Internal/FileSystem/DynamicMemoryFile.h
@@ -81,7 +81,7 @@ public:
     \brief Truncate a file to a specified length
     \param size A size, that file is going to be truncated to
     */
-    bool Truncate(int32 size) override;
+    bool Truncate(uint64 size) override;
 
 protected:
     uint64 currentPtr;
